error: implicit conversion loses integer precision: 'long' to 'int' [-Werror,-Wshorten-64-to-32]

+++ libc/test/src/sys/socket/linux/sendmsg_recvmsg_test.cpp
@@ -45,7 +45,7 @@ TEST_F(LlvmLibcSendMsgRecvMsgTest, SucceedsWithSocketPair) {
   send_message.msg_controllen = 0;
   send_message.msg_flags = 0;
 
-  ASSERT_THAT(LIBC_NAMESPACE::sendmsg(sockpair[0], &send_message, 0),
+  ASSERT_THAT((int)LIBC_NAMESPACE::sendmsg(sockpair[0], &send_message, 0),
               Succeeds(static_cast<ssize_t>(MESSAGE_LEN)));
 
   char buffer[256];
@@ -91,7 +91,7 @@ TEST_F(LlvmLibcSendMsgRecvMsgTest, SendFails) {
   send_message.msg_controllen = 0;
   send_message.msg_flags = 0;
 
-  ASSERT_THAT(LIBC_NAMESPACE::sendmsg(-1, &send_message, 0), Fails(EBADF));
+  ASSERT_THAT((int)LIBC_NAMESPACE::sendmsg(-1, &send_message, 0), Fails(EBADF));
 }
 
 TEST_F(LlvmLibcSendMsgRecvMsgTest, RecvFails) {
